updated context

+++ models/document.go
@@ -6,8 +6,8 @@ type Document struct {
 	Name       string `json:"name,omitempty"`
 	DocumentId string `json:"documentId,omitempty"`
 	FolderId   string `json:"folderId,omitempty"`
-	CreatedAt  int    `json:"created_at,omitempty"`
-	ModifiedAt int    `json:"modified_at,omitempty"`
+	CreatedAt  int64  `json:"created_at,omitempty"`
+	ModifiedAt int64  `json:"modified_at,omitempty"`
 	Owner      string `json:"owner,omitempty"`
 	ModifiedBy string `json:"modified_by,omitempty"`
 	UploadedBy string `json:"uploaded_by,omitempty"`
